+++ src/Api/Endpoints/AuthEndpoint.cs
@@ -23,6 +23,7 @@ internal sealed class AuthEndpoint : IEndpoint
             .Produces<LoginUserResponse>(StatusCodes.Status200OK)
             .ProducesProblem(StatusCodes.Status400BadRequest)
             .ProducesProblem(StatusCodes.Status401Unauthorized)
+            .ProducesProblem(StatusCodes.Status422UnprocessableEntity)
             .ProducesProblem(StatusCodes.Status500InternalServerError);
 
         endpoints.MapPost("/api/auth/logout", LogoutUser)